+++ Civi/DataProcessor/FieldOutputHandler/Calculations/CalculationFieldOutputHandler.php
@@ -7,13 +7,11 @@
 namespace Civi\DataProcessor\FieldOutputHandler\Calculations;
 
 use Civi\DataProcessor\DataSpecification\FieldSpecification;
-use Civi\DataProcessor\FieldOutputHandler\AbstractFieldOutputHandler;
+use Civi\DataProcessor\FieldOutputHandler\AbstractSimpleSortableFieldOutputHandler;
 use Civi\DataProcessor\FieldOutputHandler\FieldOutput;
-use Civi\DataProcessor\Exception\DataSourceNotFoundException;
-use Civi\DataProcessor\Exception\FieldNotFoundException;
 use CRM_Dataprocessor_ExtensionUtil as E;
 
-abstract class CalculationFieldOutputHandler extends AbstractFieldOutputHandler {
+abstract class CalculationFieldOutputHandler extends AbstractSimpleSortableFieldOutputHandler {
 
   /**
    * @var \Civi\DataProcessor\DataSpecification\FieldSpecification